// src/openapi/client/models/PlaceTranslationCreate.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { LanguagesChoices } from './LanguagesChoices';
export type PlaceTranslationCreate = {
    name: string;
    fullName?: (string | null);
    locale: LanguagesChoices;
};